Added the transfer Endpoints with support to JWT

+++ internal/dto/accounts.go
@@ -1,9 +1,9 @@
 package dto
 
 type Account struct {
-	Name       string `json:"name" validate:"required,max=8,excludesall= "`
-	Cpf        string `json:"cpf" validate:"required,len=11"`
-	Secret     string `json:"secret" validate:"required,max=50,excludesall= "`
-	Balance    int64  `json:"balance" validate:"omitempty"`
-	Created_at string `validate:"required"`
+	Name       string  `json:"name" validate:"required,max=8,excludesall= "`
+	Cpf        string  `json:"cpf" validate:"required,len=11"`
+	Secret     string  `json:"secret" validate:"required,max=50,excludesall= "`
+	Balance    float64 `json:"balance" validate:"omitempty"`
+	Created_at string  `validate:"required"`
 }
